POC instance emoji update
