+++ src/components/Emojis/List.tsx
@@ -1,9 +1,9 @@
 import { emojis } from '@components/Emojis'
 import Icon from '@components/Icon'
 import CustomText from '@components/Text'
-import { useAppDispatch } from '@root/store'
 import { useAccessibility } from '@utils/accessibility/AccessibilityManager'
-import { countInstanceEmoji } from '@utils/slices/instancesSlice'
+import { getAccountStorage, setAccountStorage } from '@utils/storage/actions'
+import { StorageAccount } from '@utils/storage/versions/account'
 import { StyleConstants } from '@utils/styles/constants'
 import layoutAnimation from '@utils/styles/layoutAnimation'
 import { useTheme } from '@utils/styles/ThemeManager'
@@ -23,7 +23,6 @@ import validUrl from 'valid-url'
 import EmojisContext from './helpers/EmojisContext'
 
 const EmojisList = () => {
-  const dispatch = useAppDispatch()
   const { reduceMotionEnabled } = useAccessibility()
   const { t } = useTranslation(['common', 'screenCompose'])
 
@@ -75,7 +74,56 @@ const EmojisList = () => {
                 key={emoji.shortcode}
                 onPress={() => {
                   addEmoji(`:${emoji.shortcode}:`)
-                  dispatch(countInstanceEmoji(emoji))
+
+                  const HALF_LIFE = 60 * 60 * 24 * 7 // 1 week
+                  const calculateScore = (
+                    emoji: StorageAccount['emojis_frequent'][number]
+                  ): number => {
+                    var seconds = (new Date().getTime() - emoji.lastUsed) / 1000
+                    var score = emoji.count + 1
+                    var order = Math.log(Math.max(score, 1)) / Math.LN10
+                    var sign = score > 0 ? 1 : score === 0 ? 0 : -1
+                    return (sign * order + seconds / HALF_LIFE) * 10
+                  }
+
+                  const currentEmojis = getAccountStorage.object('emojis_frequent')
+                  const foundEmojiIndex = currentEmojis.findIndex(
+                    e => e.emoji.shortcode === emoji.shortcode && e.emoji.url === emoji.url
+                  )
+
+                  let newEmojisSort: StorageAccount['emojis_frequent']
+                  if (foundEmojiIndex > -1) {
+                    newEmojisSort = currentEmojis
+                      .map((e, i) =>
+                        i === foundEmojiIndex
+                          ? {
+                              ...e,
+                              score: calculateScore(e),
+                              count: e.count + 1,
+                              lastUsed: new Date().getTime()
+                            }
+                          : e
+                      )
+                      .sort((a, b) => b.score - a.score)
+                  } else {
+                    newEmojisSort = currentEmojis || []
+                    const temp = {
+                      emoji,
+                      score: 0,
+                      count: 0,
+                      lastUsed: new Date().getTime()
+                    }
+                    newEmojisSort.push({
+                      ...temp,
+                      score: calculateScore(temp),
+                      count: temp.count + 1
+                    })
+                  }
+
+                  setAccountStorage(
+                    'emojis_frequent',
+                    newEmojisSort.sort((a, b) => b.score - a.score).slice(0, 20)
+                  )
                 }}
                 style={{ padding: StyleConstants.Spacing.S }}
               >
