+++ src/utils/slices/instancesSlice.ts
@@ -100,54 +100,6 @@ const instancesSlice = createSlice({
         ...instances[activeIndex].mePage,
         ...action.payload
       }
-    },
-    countInstanceEmoji: (
-      { instances },
-      action: PayloadAction<InstanceLatest['frequentEmojis'][0]['emoji']>
-    ) => {
-      const HALF_LIFE = 60 * 60 * 24 * 7 // 1 week
-      const calculateScore = (emoji: InstanceLatest['frequentEmojis'][0]): number => {
-        var seconds = (new Date().getTime() - emoji.lastUsed) / 1000
-        var score = emoji.count + 1
-        var order = Math.log(Math.max(score, 1)) / Math.LN10
-        var sign = score > 0 ? 1 : score === 0 ? 0 : -1
-        return (sign * order + seconds / HALF_LIFE) * 10
-      }
-      const activeIndex = findInstanceActive(instances)
-      const foundEmojiIndex = instances[activeIndex].frequentEmojis?.findIndex(
-        e => e.emoji.shortcode === action.payload.shortcode && e.emoji.url === action.payload.url
-      )
-      let newEmojisSort: InstanceLatest['frequentEmojis']
-      if (foundEmojiIndex > -1) {
-        newEmojisSort = instances[activeIndex].frequentEmojis
-          .map((e, i) =>
-            i === foundEmojiIndex
-              ? {
-                  ...e,
-                  score: calculateScore(e),
-                  count: e.count + 1,
-                  lastUsed: new Date().getTime()
-                }
-              : e
-          )
-          .sort((a, b) => b.score - a.score)
-      } else {
-        newEmojisSort = instances[activeIndex].frequentEmojis || []
-        const temp = {
-          emoji: action.payload,
-          score: 0,
-          count: 0,
-          lastUsed: new Date().getTime()
-        }
-        newEmojisSort.push({
-          ...temp,
-          score: calculateScore(temp),
-          count: temp.count + 1
-        })
-      }
-      instances[activeIndex].frequentEmojis = newEmojisSort
-        .sort((a, b) => b.score - a.score)
-        .slice(0, 20)
     }
   },
   extraReducers: builder => {
@@ -372,9 +324,6 @@ export const getInstanceMePage = ({ instances: { instances } }: RootState) =>
 export const getInstanceDrafts = ({ instances: { instances } }: RootState) =>
   instances[findInstanceActive(instances)]?.drafts
 
-export const getInstanceFrequentEmojis = ({ instances: { instances } }: RootState) =>
-  instances[findInstanceActive(instances)]?.frequentEmojis
-
 export const {
   updateInstanceActive,
   updateInstanceAccount,
@@ -383,8 +332,7 @@ export const {
   removeInstanceDraft,
   disableAllPushes,
   updateInstanceFollowingPage,
-  updateInstanceMePage,
-  countInstanceEmoji
+  updateInstanceMePage
 } = instancesSlice.actions
 
 export default instancesSlice.reducer
